[WIP] Add library functionality to encrypt and decrypt messages
Summary:
This commit is a work-in-progress that passes minimal tests. The goal
here is to define and implement a inter-wallet messaging protocol for
sending structured messages over telegram, or one-day a custom BCH
messaging system.  These messages can then be decoded in the wallet to
send stealth payments, encrypted messages, or other goodies.

Test Plan:
Run `tox`